python-data-engineer/week10/mini_project/requirements.txt
# Core dependencies
pandas==2.1.0
numpy==1.24.3
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# Workflow engines
apache-airflow==2.7.1
prefect==2.13.0
luigi==3.3.0

# Monitoring and metrics
prometheus-client==0.17.1
grafana-client==0.8.1
statsd==4.0.1

# Performance
dask==2023.9.0
distributed==2023.9.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# API and Web
fastapi==0.103.1
uvicorn==0.23.2 